Permit JS-disabling to show noscript elements on FF
When using uMatrix to block scripts, noscript elements are not parsed by
the browser, as uMatrix injected a CSP rule instead of disabling JS. This
breaks many pages that would work perfectly well without JS, unless the
user explicitly permits JS.

This commit is more a proof-of-concept than a complete patch, adding an
option that will cause uMatrix to disable JS in tabs whose root domain's
scripts are blocked. The option has no effect on browsers other than
Firefox (other browsers have not been tested!). This option will cause
scripts from other hosts to be blocked as well, thus it is not possible
permit third-party scripts (or individual subdomains), while not
permitting scripts to be loaded from the root domain.

Furthermore, this commit intentionally breaks localization, the new string
for this option has only be inserted for the locales 'en' and 'de'.