Add convenience option to auto-enable scripts
When browsing with a global disable-js switch, reenabling scripts for
a single page requires at least two clicks on the panel (enabling a
hostname and disabling the switch). With the new option enabled, the
second click is no longer required.

This commit intentionally breaks localization, the new string for this
option has only been inserted for the locales 'en' and 'de'.