Apply disable-js switch without reload (Firefox)
As the evaluation of the disable-js switch happened after the location
had been updated, it needed to reload the page. This causes issues with
pages that are POST'ed on a scripting barrier (that is, the POST target
host has a different disable-js switch setting than the host the POST
originates from).

This commit thus works around the issue by evaluating the disable-js
switch earlier.